Improve compatibility with some Markdown renderers
Some Markdown flavors (e.g. GitLab Favored Markdown) convert headers
to URLs replacing all characters which are not Latin and numeric
with "-". This change improves the compatibility with them through
adding an "-a" argument option which generates headers for these
Markdown flavors.

This change pulls the commander Node.js module as a dependency.